Final x et X

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: nmougino <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 16:46:12 by nmougino          #+#    #+#             */
-/*   Updated: 2016/06/09 17:33:55 by nmougino         ###   ########.fr       */
+/*   Updated: 2016/06/10 18:44:01 by nmougino         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,8 +56,8 @@ void				ft_memdel(void **ap);
 void				*ft_memmove(void *dst, const void *src, size_t len);
 void				*ft_memset(void *b, int c, size_t len);
 int					ft_min(int a, int b);
-int					ft_nbrlen(int nb);
-int					ft_nbrlenbase(int nb, int base);
+int					ft_nbrlen(long nb);
+int					ft_nbrlenbase(unsigned long nb, int base);
 int					ft_pow(int nb, int pow);
 void				ft_putchar(char c);
 void				ft_putchar_fd(char c, int fd);
